drivers: c99 inline semantics
c99 demands that inline functions are declared and implemented in the
same translation unit. extern breaks that request.
Implement the (static) inline functions in the header as is common.

This in turn requires to declare td_xenio_ctx by including the td-ctx.h
header in td-blkif.h and instead forward declare td_xenblkif.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>